backend/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
opencv-python
Pillow
python-multipart
